Promotion

+++ DongDongYee/src/main/java/domain/promotion/promotioncontroller/PromotionDeleteServlet.java
@@ -9,7 +9,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.PrintWriter;
+import javax.servlet.http.HttpSession;
 
 @WebServlet("/promotionDelete")
 public class PromotionDeleteServlet extends HttpServlet {
@@ -23,6 +23,11 @@ public class PromotionDeleteServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/html;charset=utf-8");
         request.setCharacterEncoding("UTF-8");
+        HttpSession session = request.getSession();
+        if (session.getAttribute("userID") == null) {
+            response.sendRedirect("Login.jsp");
+            return;
+        }
         promotionService.delete(Long.parseLong(request.getParameter("id")));
         response.sendRedirect("promotionlist");
     }
